Add Controller#valid_param_keys
Method just calls through to the params parser.

Add some documentation. It is again worthwhile to consider
whether the docs could be improved, particularly to show
a case in which we use the return value of the method
for keys into the params hash.

Add tests for valid_param_keys, which pass.